Sale prices of residential property in Valencia in 2018.

A dataset containing information about 79,360 real estate listings on idealista in Valencia during 2018.

Usage

data(Valencia_Sale)

Valencia_Sale

Format

An object of class `sf` (inherits from `data.frame`) with 79,360 rows and 59 variables:

ASSETID

Unique identifier of the asset

PERIOD

Date YYYYMM, indicates the quarter when the ad was extracted we used YYYY03 for the 1st Q, YYYY06 the 2nd, YYYY09 for the 3rd and YYYY12 for the 4th

PRICE

Price at idealista expressed in euros

UNITPRICE

Price in euros per square meter

CONSTRUCTEDAREA

Home built area in sq.m

ROOMNUMBER

Number of bedrooms

BATHNUMBER

Number of bathrooms

HASTERRACE

Dummy variable for terrace (takes 1 if there is a terrace, 0 otherwise)

HASLIFT

Dummy variable for lift (takes 1 if there is a lift in the building, 0 otherwise)

HASAIRCONDITIONING

Dummy variable for AA (takes 1 if there is a AA, 0 otherwise)

AMENITYID

Indicates the amenities included (1 - no furniture, no kitchen amenities, 2 - kitchen amenities, no furniture, 3 - kitchen amenities, furniture)

HASPARKINGSPACE

Dummy variable for parking (takes 1 if parking is included in the Ad, 0 otherwise)

ISPARKINGSPACEINCLUDEDINPRICE

Dummy variable for parking (takes 1 if parking is included in the Ad, 0 otherwise)

PARKINGSPACEPRICE

Price of parking Space in euros

HASNORTHORIENTATION

Dummy variable for orientation (takes 1 if orientation is North in the Ad, 0 otherwise) - *Important note:* orientation features are not orthogonal features, a house oriented to the north can be also oriented to the east

HASSOUTHORIENTATION

Dummy variable for orientation (takes 1 if orientation is South in the Ad, 0 otherwise) - Important note: orientation features are not orthogonal features, a house oriented to the north can be also oriented to the east

HASEASTORIENTATION

Dummy variable for orientation (takes 1 if orientation is East in the Ad, 0 otherwise) - Important note: orientation features are not orthogonal features, a house oriented to the north can be also oriented to the east

HASWESTORIENTATION

Dummy variable for orientation (takes 1 if orientation is West in the Ad, 0 otherwise) - Important note: orientation features are not orthogonal features, a house oriented to the north can be also oriented to the east

HASBOXROOM

Dummy variable for boxroom (takes 1 if boxroom is included in the Ad, 0 otherwise)

HASWARDROBE

Dummy variable for wardrobe (takes 1 if wardrobe is included in the Ad, 0 otherwise)

HASSWIMMINGPOOL

Dummy variable for swimming pool (takes 1 if swimming pool is included in the Ad, 0 otherwise)

HASDOORMAN

Dummy variable for doorman (takes 1 if there is a doorman in the building, 0 otherwise)

HASGARDEN

Dummy variable for garden (takes 1 if there is a garden in the building, 0 otherwise)

ISDUPLEX

Dummy variable for duplex (takes 1 if it is a duplex, 0 otherwise)

ISSTUDIO

Dummy variable for bachelor apartment (studio in Spanish) (takes 1 if it is a single-person apartment, 0 otherwise)

ISINTOPFLOOR

Dummy variable indicating if the apartment is located on the top floor (takes 1 on the top floor 0 otherwise)

CONSTRUCTIONYEAR

Construction year (source: advertiser)

FLOORCLEAN

Indicates flat floornumber starting from the 0 value for groung floor (source: advertiser)

FLATLOCATIONID

Indicates the kind of views the flat has (1 - external, 2 - internal)

CADCONSTRUCTIONYEAR

Construction year as of cadastral source (source: cadastre), note this figure can differ from the one given by the advertiser

CADMAXBUILDINGFLOOR

Max building floor (source: cadastre)

CADDWELLINGCOUNT

Dwelling count in the building (source: cadastre)

CADASTRALQUALITYID

Cadastral quality (source: cadastre)

BUILTTYPEID_1

Dummy value for flat condition: 1 new development 0 otherwise (source: advertiser))

BUILTTYPEID_2

Dummy value for flat condition: 1 second hand to be restored 0 otherwise (source: advertiser))

BUILTTYPEID_3

Dummy value for flat condition: 1 second hand in good condition 0 otherwise (source: advertiser))

DISTANCE_TO_CITY_CENTER

Distance to center of city in km

DISTANCE_TO_METRO

Distance to the nearest subway entrance in km

DISTANCE_TO_BLASCO

Distance to Blasco Ibáñez Avenue in km; Blasco Ibañez is a major street in the city

geometry

geometry of simple features

An object of class sf (inherits from data.frame) with 33622 rows and 42 columns.

Source

Idealista (https://www.idealista.com/)

Examples

 library(idealista18)
 data(Valencia_Sale)
 summary(Valencia_Sale$PRICE)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   20000   99000  151000  199678  234000 2772000