Total number of properties by district in 2018.

A dataset containing information about the number of properties by district in Barcelona, Madrid, and Valencia during 2018 according to the cadastre. Districts are collections of neighborhoods.

Usage

data(properties_by_district)

properties_by_district

Format

An object of class `sf` (inherits from `data.frame`) with 50 rows and 5 variables:

District

Number of district

N_CADASTRE

Date YYYYMM, indicates the quarter when the ad was extracted we used YYYY03 for the 1st Q, YYYY06 the 2nd, YYYY09 for the 3rd and YYYY12 for the 4th

N

Asking price for the ad at idealista expressed in euros

CITY

Price in euros per square meter

geometry

geometry of simple features in latitude and longitude

An object of class sf (inherits from tbl_df, tbl, data.frame) with 47 rows and 5 columns.

Source

Idealista (https://www.idealista.com/)

Examples

 library(idealista18)
 data(properties_by_district)
 summary(properties_by_district$N)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>     728    2358    3494    4036    5084   13936